+++ Render/src/VertexBuffer.cpp
@@ -4,7 +4,7 @@
 
 namespace Renderer {
 
-  VertexBuffer::VertexBuffer(const void* data, uint32_t size) 
+  VertexBuffer::VertexBuffer(const void* data, size_t size)
     : _size(size) {
     glGenBuffers(1, &_buffer_id);
     glBindBuffer(GL_ARRAY_BUFFER, _buffer_id);
